Final Checking Numbers
I'm unsure as to what the .count() method was suppose to be used for in
this exercise. It immediately exits the loop if it finds a single match.
How can we use that to count?

Also creating the random number in a seperate method and looping through
the list seems a lot more intense than just checking if the random
number is equal to the one's populating the list.